Add a message on installation
A message is added for the install() function, such that the user is notified of the new updated version available on GitHub. This message occurs just once after the package is loaded. Also the user is prompted so that they can optionally quit the installation step.

General refactoring was carried out as the install() function was becoming rather unwieldy. Internal functions were moved to package environment. NAMEPACE was updated to reflect the use of purrr::iwalk in place of base::Map.