Try to fix R CMD check
There is an issue with running `R CMD check` via GitHub Actions. The problem seems to be with the inability to download and install RQDA at that level (because the package has been archived, which is one of the reasons this package was developed in the first place).

The change that was applied is in the script for GA. I noticed that the function `install_deps` from the 'remotes' package is used to install the package dependencies, including RQDA. In the package, RQDA is a 'Suggested' package, so I have tried to modify the dependencies `install_deps` will look for. According to its help file, setting the `dependencies` argument to `NA` will exclude the installation of packages listed in the `Suggests` field. So, I'm committing and pushing to GitHub to see whether this helps in any way.