Reverse parts of last change
In the last commit, I introduced code for the script that explicitly enables the installation of RQDA and its dependencies (archived versions). The snag I discovered is the absence of GTK+, which is a system requirement for the installation of RGtk2. Gtk+ cannot be installed via the regular R package installation workflow. So, it's either I devise how to install it via the script. It's something I might revisit again.