Enable installation of RGtk2 archive
RGtk2 can now be installed from the CRAN archive. The problem of the
system requirements, namely the Gtk+ distribution with all its components
was solved by automatic download and placement on System PATH to enable
compilation of the package. Also, the distribution is copied into the
package for successful loading of its namespace. So far, this has only
been tested on Windows.