Update the logic for source installation
This specifically refers to the inability to build cairoDevice when GTK is not installed to C:/ and environment variable GTK_PATH is not set. Initially, the package allowed binary installation of cairoDevice. This is no longer possible, so there is a need to ensure that its dependencies are satisfied. Discovered by issue filed by @Joseph1Otoo.