Streamline installation process
The installation is now to include the base requirements (R 4.1.3 and
Rtools40). Their presence are checked for i.e. whether they are
installed via the Windows Registry entries. Then, some of the redundant
checks that were previously being done in the 'install' function have
now been delegated to the 'install_rgtk2_and_deps' function. The
installation of the R packages is now handled more smoothly.