Download and install R and Rtools
The appropriate versions of R (4.1.3) and Rtools40 are now downloadable and installable as part of the `install()` functionality, when  used in non-compatible versions of R. To proceed with the installatioin of the other RQDA dependencies, the user is forced to exit the session once these 2 are installed. Further actions have to be carried out in an active R session of the right version.