Uprade GitHub Actions workflow to V2
A number of changes were made to try to overcome the issue with checking the package in R 4.0. At this point, it checks out well on the local machine. But later, I discovered that the workflow has actually been updated, so were pushing it to see if there's a difference or we need to further tweak it.